Many to many connections
Currently graph connections are always one output to many inputs. This
is appropriate for most "data type values", but not always correct. The
motivating example for a place where it is incorrect is "control flow",
where many nodes might "output" control flow to the same node's control
flow input.

This changes that so that it is configurable via data-type whether nodes
are one-many, many-one, many-many, or even one-one with the concept of
splittable data-types (data types that can be copied from one output to
many inputs) and mergeable data-types (data types that can accept many
outputs into a single input).